GET Attribute, fixes for Simplified
